+++ mobile/android/android-components/components/browser/tabstray/src/test/java/mozilla/components/browser/tabstray/DefaultTabViewHolderTest.kt
@@ -24,7 +24,6 @@ import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.mockito.Mockito.doReturn
 import org.mockito.Mockito.never
 import org.mockito.Mockito.verify
 
@@ -37,13 +36,13 @@ class DefaultTabViewHolderTest {
         val titleView = view.findViewById<TextView>(R.id.mozac_browser_tabstray_title)
         val urlView = view.findViewById<TextView>(R.id.mozac_browser_tabstray_url)
 
-        val holder = DefaultTabViewHolder(view, mockTabsTrayWithStyles())
+        val holder = DefaultTabViewHolder(view)
 
         assertEquals("", titleView.text)
 
         val session = Tab("a", "https://www.mozilla.org")
 
-        holder.bind(session, isSelected = false, observable = mock())
+        holder.bind(session, isSelected = false, styling = mock(), observable = mock())
 
         assertEquals("https://www.mozilla.org", titleView.text)
         assertEquals("www.mozilla.org", urlView.text)
@@ -53,10 +52,10 @@ class DefaultTabViewHolderTest {
     fun `URL text is set to tab URL when exception is thrown`() {
         val view = LayoutInflater.from(testContext).inflate(R.layout.mozac_browser_tabstray_item, null)
         val urlView = view.findViewById<TextView>(R.id.mozac_browser_tabstray_url)
-        val holder = DefaultTabViewHolder(view, mockTabsTrayWithStyles())
+        val holder = DefaultTabViewHolder(view)
         val session = Tab("a", "about:home")
 
-        holder.bind(session, isSelected = false, observable = mock())
+        holder.bind(session, isSelected = false, styling = mock(), observable = mock())
 
         assertEquals("about:home", urlView.text)
     }
@@ -69,10 +68,10 @@ class DefaultTabViewHolderTest {
         }
 
         val view = LayoutInflater.from(testContext).inflate(R.layout.mozac_browser_tabstray_item, null)
-        val holder = DefaultTabViewHolder(view, mockTabsTrayWithStyles())
+        val holder = DefaultTabViewHolder(view)
 
         val session = Tab("a", "https://www.mozilla.org")
-        holder.bind(session, isSelected = false, observable = registry)
+        holder.bind(session, isSelected = false, styling = mock(), observable = registry)
 
         view.performClick()
 
@@ -87,10 +86,10 @@ class DefaultTabViewHolderTest {
         }
 
         val view = LayoutInflater.from(testContext).inflate(R.layout.mozac_browser_tabstray_item, null)
-        val holder = DefaultTabViewHolder(view, mockTabsTrayWithStyles())
+        val holder = DefaultTabViewHolder(view)
 
         val session = Tab("a", "https://www.mozilla.org")
-        holder.bind(session, isSelected = true, observable = registry)
+        holder.bind(session, isSelected = true, styling = mock(), observable = registry)
 
         view.findViewById<View>(R.id.mozac_browser_tabstray_close).performClick()
 
@@ -105,13 +104,13 @@ class DefaultTabViewHolderTest {
         }
 
         val view = LayoutInflater.from(testContext).inflate(R.layout.mozac_browser_tabstray_item, null)
-        val holder = DefaultTabViewHolder(view, mockTabsTrayWithStyles())
+        val holder = DefaultTabViewHolder(view)
 
         val session = Tab("a", "https://www.mozilla.org")
         val titleView = holder.itemView.findViewById<TextView>(R.id.mozac_browser_tabstray_title)
         val urlView = view.findViewById<TextView>(R.id.mozac_browser_tabstray_url)
 
-        holder.bind(session, isSelected = true, observable = registry)
+        holder.bind(session, isSelected = true, styling = mock(), observable = registry)
 
         assertEquals(session.url, titleView.text)
         assertEquals("www.mozilla.org", urlView.text)
@@ -125,13 +124,13 @@ class DefaultTabViewHolderTest {
         }
 
         val view = LayoutInflater.from(testContext).inflate(R.layout.mozac_browser_tabstray_item, null)
-        val holder = DefaultTabViewHolder(view, mockTabsTrayWithStyles())
+        val holder = DefaultTabViewHolder(view)
 
         val session = Tab("a", "https://www.mozilla.org", title = "Mozilla Firefox")
         val titleView = holder.itemView.findViewById<TextView>(R.id.mozac_browser_tabstray_title)
         val urlView = view.findViewById<TextView>(R.id.mozac_browser_tabstray_url)
 
-        holder.bind(session, isSelected = true, observable = registry)
+        holder.bind(session, isSelected = true, styling = mock(), observable = registry)
 
         assertEquals("Mozilla Firefox", titleView.text)
         assertEquals("www.mozilla.org", urlView.text)
@@ -142,13 +141,13 @@ class DefaultTabViewHolderTest {
         val view = LayoutInflater.from(testContext).inflate(R.layout.mozac_browser_tabstray_item, null)
         val thumbnailView = view.findViewById<ImageView>(R.id.mozac_browser_tabstray_thumbnail)
 
-        val holder = DefaultTabViewHolder(view, mockTabsTrayWithStyles())
+        val holder = DefaultTabViewHolder(view)
         assertEquals(null, thumbnailView.drawable)
 
         val emptyBitmap = Bitmap.createBitmap(40, 40, Bitmap.Config.ARGB_8888)
         val session = Tab("a", "https://www.mozilla.org", thumbnail = emptyBitmap)
 
-        holder.bind(session, isSelected = false, observable = mock())
+        holder.bind(session, isSelected = false, styling = mock(), observable = mock())
         assertTrue(thumbnailView.drawable != null)
     }
 
@@ -156,15 +155,15 @@ class DefaultTabViewHolderTest {
     fun `thumbnail is set from loader`() {
         val view = LayoutInflater.from(testContext).inflate(R.layout.mozac_browser_tabstray_item, null)
         val loader: ImageLoader = mock()
-        val viewHolder = DefaultTabViewHolder(view, mockTabsTrayWithStyles(), loader)
+        val viewHolder = DefaultTabViewHolder(view, loader)
         val tabWithThumbnail = Tab("123", "https://example.com", thumbnail = mock())
         val tab = Tab("123", "https://example.com")
 
-        viewHolder.bind(tabWithThumbnail, false, mock())
+        viewHolder.bind(tabWithThumbnail, false, mock(), mock())
 
         verify(loader, never()).loadIntoView(any(), eq(ImageLoadRequest("123", 100)), nullable(), nullable())
 
-        viewHolder.bind(tab, false, mock())
+        viewHolder.bind(tab, false, mock(), mock())
 
         verify(loader).loadIntoView(any(), eq(ImageLoadRequest("123", 100)), nullable(), nullable())
     }
@@ -172,26 +171,15 @@ class DefaultTabViewHolderTest {
     @Test
     fun `thumbnailView does not change when there is no cache or new thumbnail`() {
         val view = LayoutInflater.from(testContext).inflate(R.layout.mozac_browser_tabstray_item, null)
-        val viewHolder = DefaultTabViewHolder(view, mockTabsTrayWithStyles())
+        val viewHolder = DefaultTabViewHolder(view)
         val tab = Tab("123", "https://example.com")
         val thumbnailView = view.findViewById<ImageView>(R.id.mozac_browser_tabstray_thumbnail)
 
         thumbnailView.setImageBitmap(mock())
         val drawable = thumbnailView.drawable
 
-        viewHolder.bind(tab, false, mock())
+        viewHolder.bind(tab, false, mock(), mock())
 
         assertEquals(drawable, thumbnailView.drawable)
-    }
-
-    companion object {
-        fun mockTabsTrayWithStyles(): BrowserTabsTray {
-            val styles: TabsTrayStyling = mock()
-
-            val tabsTray: BrowserTabsTray = mock()
-            doReturn(styles).`when`(tabsTray).styling
-
-            return tabsTray
-        }
     }
 }